Do not query pod tasks.
