+++ src/test/scala/mesosphere/marathon/integration/UpgradeIntegrationTest.scala
@@ -178,8 +178,7 @@ class UpgradeIntegrationTest extends AkkaIntegrationTest with MesosClusterTest w
         mountPath = containerPath,
         cmd = s"""echo "data" > $containerPath/data && while test -e foo; do sleep 5; done""")
       marathon16322.client.createPodV2(resident_pod_16322) should be(Created)
-      eventually { marathon16322 should have(runningTasksFor(resident_pod_16322.id, 1)) }
-      val originalPersistentPodTasks = marathon16322.client.tasks(resident_pod_16322.id).value
+      eventually { marathon16322.client.status(resident_pod_16322.id) should be(Stable) }
 
       Then("All apps from 1.4.9 and 1.5.6 are still running")
       marathon16322.client.tasks(app_149.id.toPath).value should contain theSameElementsAs (originalApp149Tasks)
@@ -203,8 +202,7 @@ class UpgradeIntegrationTest extends AkkaIntegrationTest with MesosClusterTest w
       marathonCurrent.client.tasks(app_156_fail.id.toPath).value should not contain theSameElementsAs(originalApp156FailedTasks)
 
       And("All pods from 1.6.322 are still running")
-      eventually { marathonCurrent should have(runningTasksFor(resident_pod_16322.id, 1)) }
-      marathon16322.client.tasks(resident_pod_16322.id).value should contain theSameElementsAs (originalPersistentPodTasks)
+      eventually { marathonCurrent.client.status(resident_pod_16322.id) should be(Stable) }
 
       marathonCurrent.close()
     }
